Echo/requirements.txt
gradio>=4.44.0,<5.0.0
numpy>=1.26.0,<3.0.0
opencv-python>=4.10.0,<5.0.0
Pillow>=10.4.0,<12.0.0
